refactor: moving the managementgroups client
